refactor(storage): move upload progress parser
